Rewrite package to use newer Proj.jl and be a modern package
Previously, we relied on on Proj4.jl, which is no longer developed.
That version of the Proj wrapper used a different method to calculate
the conversion between the National Grid and WGS84 coordinates.  Newer
versions of Proj know to automatically use a set of grid points to do
the conversion.  This requires us to create an artifact (which has to be
an optionally compressed tar file) containing these grid points, because
Proj does not come with them.  Therefore, the new version of this
package ships the required artifacts and performs a more accurate
conversion than before.  It also becomes a modern Julia package with a
`Project.toml` file and so on.

Along the way, we can drop the requirement on Formatting.jl and use the
features in the Printf standard library since Juliva v1.6.  This means
v1.6 is now our minimum version.

Because this therefore breaks semver, we also take the opportunity to
slightly change the API.  Overall, therefore, the breaking changes are:

- Change the lon-lat to BNG conversions so they are more accurate
- Make some of the arguments to `gridref` keyword arguments

The major non-breaking change is:
- Update Julia minimum version to v1.6